chore: add baseQuery handling authorization header. Update countriesApi and eventsApi (DRY principle)
