+++ src/redux/services/countriesApi/countriesApi.ts
@@ -1,20 +1,12 @@
-import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
+import { createApi } from "@reduxjs/toolkit/query/react";
 import ROUTES from "@/lib/routes/routes";
 import { CountriesEntity } from "./countriesApi.types";
+import { baseQuery } from "@/redux/services/baseQuery";
 
 // create and export service
 export const countriesApi = createApi({
   reducerPath: "countriesApi",
-  baseQuery: fetchBaseQuery({
-    baseUrl: "",
-    prepareHeaders: (headers) => {
-      const token = localStorage.getItem("token");
-      if (token) {
-        headers.set("Authorization", `Bearer ${token}`);
-      }
-      return headers;
-    },
-  }),
+  baseQuery: baseQuery,
   endpoints: (builder) => ({
     getCountries: builder.query<CountriesEntity, void>({ query: () => ROUTES.ALL_COUNTRIES }),
   }),
