+++ src/redux/services/eventsApi/eventsApi.ts
@@ -1,20 +1,12 @@
-import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
+import { createApi } from "@reduxjs/toolkit/query/react";
 import { UpcomingEventsEntity, UpcomingEventsQuery } from "./eventsApi.types";
 import ROUTES from "@/lib/routes/routes";
+import { baseQuery } from "@/redux/services/baseQuery";
 
 // service
 export const eventsApi = createApi({
   reducerPath: "eventsApi",
-  baseQuery: fetchBaseQuery({
-    baseUrl: "",
-    prepareHeaders: (headers) => {
-      const token = localStorage.getItem("token");
-      if (token) {
-        headers.set("Authorization", `Bearer ${token}`);
-      }
-      return headers;
-    },
-  }),
+  baseQuery: baseQuery,
   endpoints: (builder) => ({
     getUpcomingEvents: builder.query<UpcomingEventsEntity, UpcomingEventsQuery>({
       query: ({ page, onlineOnly, country }) =>
